♻️ Organize code
